Modified cut-in to LaneOffset instead of LaneChange. EnvSimMini updated accordingly.
Former-commit-id: df576f9eaa1c1ba3a6efe815b3b759db359a5fd0

+++ EnvironmentSimulator/ScenarioEngineMini/Actions/Action.hpp
@@ -46,7 +46,7 @@ private:
 	double targetValue;
 	double time;
 	double f;
-	std::vector<int> initialOffsets;
+	std::vector<double> initialOffsets;
 
 
 	// Speed
